pyproject.toml
[project]
name = "django-auditlog-example"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Ryuji TSUTSUI", email = "ryu22e@gmail.com" }
]
dependencies = [
    "django>=5.0, <5.1",
    "django-auditlog>=2.3.0",
    "djangorestframework>=3.14.0",
    "markdown>=3.5.2",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "ruff==0.3.2",
    "pre-commit>=3.6.2",
    "pytest-django>=4.8.0",
    "factory_boy>=3.3.0",
    "pytest-factoryboy>=2.7.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/django_auditlog_example"]

[tool.pytest.ini_options]
python_files = ["test_*.py", "*_test.py", "tests.py"]
DJANGO_SETTINGS_MODULE = "django_auditlog_example.settings"
FAIL_INVALID_TEMPLATE_VARS = true

[tool.ruff.lint.isort]
combine-as-imports = true
